+++ src/main/java/blackjack/domain/Denomination.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 public enum Denomination {
     ACE(1),
