+++ src/test/java/blackjack/BlackjackTest.java
@@ -2,8 +2,14 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.*;
 
-import blackjack.user.Dealer;
-import blackjack.user.Player;
+import blackjack.domain.Blackjack;
+import blackjack.domain.Card;
+import blackjack.domain.Deck;
+import blackjack.domain.Denomination;
+import blackjack.domain.Players;
+import blackjack.domain.Suit;
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Player;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
