+++ src/test/java/blackjack/CardTest.java
@@ -2,6 +2,9 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.*;
 
+import blackjack.domain.Card;
+import blackjack.domain.Denomination;
+import blackjack.domain.Suit;
 import org.junit.jupiter.api.Test;
 
 public class CardTest {
